more refactoring, also implemented Cathals latest changes
