+++ game-service/src/main/java/com/lms/gameservice/controller/GameController.java
@@ -5,14 +5,9 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.TemporalAdjusters;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,18 +17,17 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.client.RestTemplate;
 
 import com.lms.gameservice.gamerequest.GameRequestDTO;
 import com.lms.gameservice.model.Game;
 import com.lms.gameservice.model.Player;
 import com.lms.gameservice.model.Results;
 import com.lms.gameservice.repository.GameRepository;
-import com.lms.gameservice.repository.PlayerRepository;
 import com.lms.gameservice.repository.ResultsRepository;
 import com.lms.gameservice.service.AuthService;
 import com.lms.gameservice.service.GameService;
 import com.lms.gameservice.service.InformationServiceClient;
+import com.lms.gameservice.service.NotificationServiceClient;
 import com.lms.gameservice.service.PlayerService;
 
 import net.minidev.json.JSONObject;
@@ -47,27 +41,25 @@ import com.lms.gameservice.matches.MatchesDTO;
 public class GameController {
     private final GameService gameService;
     private final PlayerService playerService;
-    private final PlayerRepository playerRepository;
     private final GameRepository gameRepository;
-    private final RestTemplate restTemplate;
 
     //for results testing
     private final InformationServiceClient info;
     private final ResultsRepository resultsRepository;
-    private static final Logger logger = LogManager.getLogger(GameController.class);
+
+    private final NotificationServiceClient noti;
 
     @Autowired
     private AuthService authService;
 
     @Autowired
-    public GameController(GameService gameService, PlayerService playerService, PlayerRepository playerRepository, GameRepository gameRepository, RestTemplate restTemplate, InformationServiceClient info, ResultsRepository resultsRepository) {
+    public GameController(GameService gameService, PlayerService playerService, GameRepository gameRepository, InformationServiceClient info, ResultsRepository resultsRepository, NotificationServiceClient noti) {
         this.gameService = gameService;
         this.playerService = playerService;
-        this.playerRepository = playerRepository;
         this.gameRepository = gameRepository;
-        this.restTemplate = restTemplate;
         this.info = info;
         this.resultsRepository = resultsRepository;
+        this.noti = noti;
     }
 
     /**
@@ -85,15 +77,15 @@ public class GameController {
         try {
             String uid = authService.validateToken(authorisationHeader);
 
-            String userID = extractUidFromMessage(uid);
+            String userID = noti.extractUidFromMessage(uid);
 
-            String userEmail = getUserEmailByUid(userID);
+            String userEmail = noti.getUserEmailByUid(userID);
 
             // Step 2: Create the game with the start date
             Game game = gameService.createGame(gameRequest.getName(), gameRequest.getEntryFee(),
                     gameRequest.getWeeksTillStartDate(), uid);
 
-            sendGameCreationNotification(userEmail, "game_created", gameRequest.getName(), gameRequest.getWeeksTillStartDate(), gameRequest.getEntryFee().doubleValue());
+            noti.sendGameCreationNotification(userEmail, "game_created", gameRequest.getName(), gameRequest.getWeeksTillStartDate(), gameRequest.getEntryFee().doubleValue());
             System.out.print(userEmail);
             return ResponseEntity.ok(game);
 
@@ -131,7 +123,7 @@ public class GameController {
             String[] splits = msg.split("Access granted for user: ");
             String uid = splits[1];
 
-            String userEmail = getUserEmailByUid(uid);
+            String userEmail = noti.getUserEmailByUid(uid);
 
             Game game = gameRepository.findById(gameId)
                 .orElseThrow(() -> new IllegalArgumentException("Game not found"));
@@ -145,7 +137,7 @@ public class GameController {
             boolean joinedSuccessfully = gameService.joinGame(gameId, uid, authorisationHeader);
             if (joinedSuccessfully) {
                 System.out.println("User has joined game " +  gameId + " successfully.");
-                sendGameJoinedNotification(userEmail, "game_joined", gameName, entryFee);
+                noti.sendGameJoinedNotification(userEmail, "game_joined", gameName, entryFee);
                 return ResponseEntity.ok("User joined game " + gameId);
             } else {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Unable to join the game.");
@@ -304,64 +296,6 @@ public class GameController {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
                     .body("Error verifying the player: " + e.getMessage());
         }
-    }
-
-    public void sendGameCreationNotification(String recipient, String type, String gameName, int weeksTillStartDate, double entryFee) {
-        String notificationUrl = "http://localhost:8085/api/notifications/send";
-        Map<String, String> notificationData = new HashMap<>();
-        notificationData.put("recipient", recipient);
-        notificationData.put("type", type);
-        notificationData.put("gameName", gameName);
-        notificationData.put("weeksTillStartDate", String.valueOf(weeksTillStartDate));
-        notificationData.put("entryFee", String.valueOf(entryFee));
-        try {
-            restTemplate.postForEntity(notificationUrl, notificationData, String.class);
-            logger.info("Notification request sent for type: {}", type);
-        } catch (Exception e) {
-            logger.error("Failed to send notification request for type {}: {}", type, e.getMessage());
-        }
-    }
-
-    public void sendGameJoinedNotification(String recipient, String type, String gameName, double entryFee) {
-        String notificationUrl = "http://localhost:8085/api/notifications/send";
-        Map<String, String> notificationData = new HashMap<>();
-        notificationData.put("recipient", recipient);
-        notificationData.put("type", type);
-        notificationData.put("gameName", gameName);
-        notificationData.put("entryFee", String.valueOf(entryFee));
-        try {
-            restTemplate.postForEntity(notificationUrl, notificationData, String.class);
-            logger.info("Notification request sent for type: {}", type);
-        } catch (Exception e) {
-            logger.error("Failed to send notification request for type {}: {}", type, e.getMessage());
-        }
-    }
-
-    private String getUserEmailByUid(String uid) {
-        // Call the UserController's endpoint to get the email
-        String url = "http://localhost:8080/api/users/" + uid + "/email";
-
-        try {
-            ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.GET, null, String.class);
-            if (response.getStatusCode() == HttpStatus.OK) {
-                return response.getBody(); // Return the email
-            } else {
-                return null; // Handle the case when the user is not found
-            }
-        } catch (Exception e) {
-            // Handle errors
-            return null;
-        }
-    }
-    
-    private String extractUidFromMessage(String message) {
-        // Check if the message starts with "Access granted for user: "
-        String prefix = "Access granted for user: ";
-        if (message != null && message.startsWith(prefix)) {
-            return message.substring(prefix.length());  // Extract the UID
-        }
-        return null; // Handle errors
-  
     }
     
     /**
