+++ game-service/src/main/java/com/lms/gameservice/service/NotificationServiceClient.java
@@ -37,7 +37,7 @@ public class NotificationServiceClient {
         String playerStatus,
         String playerTeamPick
     ) {
-        String notificationUrl = "http://localhost:8085/api/notifications/send";
+        String notificationUrl = "http://notification-service:8085/api/notifications/send";
 
         // Build the notification data as a map of individual fields
         Map<String, String> notificationData = new HashMap<>();
@@ -59,11 +59,41 @@ public class NotificationServiceClient {
         }
     }
 
+    public void sendGameCreationNotification(String recipient, String type, String gameName, int weeksTillStartDate, double entryFee) {
+        String notificationUrl = "http://notification-service:8085/api/notifications/send";
+        Map<String, String> notificationData = new HashMap<>();
+        notificationData.put("recipient", recipient);
+        notificationData.put("type", type);
+        notificationData.put("gameName", gameName);
+        notificationData.put("weeksTillStartDate", String.valueOf(weeksTillStartDate));
+        notificationData.put("entryFee", String.valueOf(entryFee));
+        try {
+            restTemplate.postForEntity(notificationUrl, notificationData, String.class);
+            logger.info("Notification request sent for type: {}", type);
+        } catch (Exception e) {
+            logger.error("Failed to send notification request for type {}: {}", type, e.getMessage());
+        }
+    }
+
+    public void sendGameJoinedNotification(String recipient, String type, String gameName, double entryFee) {
+        String notificationUrl = "http://notification-service:8085/api/notifications/send";
+        Map<String, String> notificationData = new HashMap<>();
+        notificationData.put("recipient", recipient);
+        notificationData.put("type", type);
+        notificationData.put("gameName", gameName);
+        notificationData.put("entryFee", String.valueOf(entryFee));
+        try {
+            restTemplate.postForEntity(notificationUrl, notificationData, String.class);
+            logger.info("Notification request sent for type: {}", type);
+        } catch (Exception e) {
+            logger.error("Failed to send notification request for type {}: {}", type, e.getMessage());
+        }
+    }
     
 
     public String getUserEmailByUid(String uid) {
         // Call the UserController's endpoint to get the email
-        String url = "http://localhost:8080/api/users/" + uid + "/email";
+        String url = "http://user-service:8080/api/users/" + uid + "/email";
 
         try {
             ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.GET, null, String.class);
@@ -76,6 +106,16 @@ public class NotificationServiceClient {
             // Handle errors
             return null;
         }
+    }
+
+    public String extractUidFromMessage(String message) {
+        // Check if the message starts with "Access granted for user: "
+        String prefix = "Access granted for user: ";
+        if (message != null && message.startsWith(prefix)) {
+            return message.substring(prefix.length());  // Extract the UID
+        }
+        return null; // Handle errors
+  
     }
 
 }
